flow_api/requirements.txt
aiocache[redis]==0.12.2 ; python_version >= "3.10" and python_version < "4.0"
aiodns==3.2.0 ; (sys_platform == "linux" or sys_platform == "darwin") and python_version >= "3.10" and python_version < "4.0"
aiofiles==23.2.1 ; python_version >= "3.10" and python_version < "4.0"
aiogram==3.6.0 ; python_version >= "3.10" and python_version < "4.0"
aiohttp==3.9.5 ; python_version >= "3.10" and python_version < "4.0"
aiohttp[speedups]==3.9.5 ; python_version >= "3.10" and python_version < "4.0"
aioredis==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
aiosqlite==0.17.0 ; python_version >= "3.10" and python_version < "4.0"
alembic==1.13.1 ; python_version >= "3.10" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==4.3.0 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.10" and python_version < "4.0"
asyncpg==0.29.0 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.10" and python_version < "4.0"
babel==2.15.0 ; python_version >= "3.10" and python_version < "4.0"
bcrypt==4.1.3 ; python_version >= "3.10" and python_version < "4.0"
boto3==1.34.113 ; python_version >= "3.10" and python_version < "4.0"
botocore==1.34.113 ; python_version >= "3.10" and python_version < "4.0"
brotli==1.1.0 ; platform_python_implementation == "CPython" and python_version >= "3.10" and python_version < "4.0"
brotlicffi==1.1.0.0 ; platform_python_implementation != "CPython" and python_version >= "3.10" and python_version < "4.0"
cachetools==5.3.3 ; python_version >= "3.10" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
cryptography==42.0.7 ; python_version >= "3.10" and python_version < "4.0"
ecs-logging==2.1.0 ; python_version >= "3.10" and python_version < "4"
elastic-apm==6.22.2 ; python_version >= "3.10" and python_version < "4"
exceptiongroup==1.2.1 ; python_version >= "3.10" and python_version < "3.11"
fastapi-admin==1.0.4 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.109.2 ; python_version >= "3.10" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.10" and python_version < "4.0"
geoalchemy2==0.15.1 ; python_version >= "3.10" and python_version < "4.0"
greenlet==3.0.3 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.6.1 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.7 ; python_version >= "3.10" and python_version < "4.0"
iso8601==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.10" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
loguru==0.7.2 ; python_version >= "3.10" and python_version < "4.0"
magic-filter==1.0.12 ; python_version >= "3.10" and python_version < "4.0"
mako==1.3.5 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.10" and python_version < "4.0"
orjson==3.10.3 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.10" and python_version < "4.0"
pendulum==3.0.0 ; python_version >= "3.10" and python_version < "4.0"
pillow==10.3.0 ; python_version >= "3.10" and python_version < "4.0"
prometheus-client==0.20.0 ; python_version >= "3.10" and python_version < "4.0"
pycares==4.4.0 ; (sys_platform == "linux" or sys_platform == "darwin") and python_version >= "3.10" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0"
pydantic-core==2.18.2 ; python_version >= "3.10" and python_version < "4.0"
pydantic-settings==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.7.1 ; python_version >= "3.10" and python_version < "4.0"
pyjwt==2.8.0 ; python_version >= "3.10" and python_version < "4.0"
pypika-tortoise==0.1.6 ; python_version >= "3.10" and python_version < "4.0"
pypng==0.20220715.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
python-multipart==0.0.9 ; python_version >= "3.10" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
pyzbar==0.1.9 ; python_version >= "3.10" and python_version < "4.0"
qrcode==7.4.2 ; python_version >= "3.10" and python_version < "4.0"
redis==5.0.4 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.2 ; python_version >= "3.10" and python_version < "4.0"
s3transfer==0.10.1 ; python_version >= "3.10" and python_version < "4.0"
sentry-sdk[loguru]==1.45.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy-utils==0.41.2 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==2.0.30 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.36.3 ; python_version >= "3.10" and python_version < "4.0"
tortoise-orm==0.21.2 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.29.0 ; python_version >= "3.10" and python_version < "4.0"
uvicorn[standard]==0.29.0 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.19.0 ; python_version >= "3.10" and python_version < "4.0"
watchfiles==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
websockets==12.0 ; python_version >= "3.10" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
wrapt==1.16.0 ; python_version >= "3.10" and python_version < "4"
yarl==1.12.1; python_version >= "3.10" and python_version < "4.0"
beautifulsoup4==4.12.3
web3==6.17.2
